+++ inc/notificationtarget.class.php
@@ -753,8 +753,11 @@ class NotificationTarget extends CommonDBChild {
     */
    final public function getDistinctUserCriteria() {
       return [
-         'SELECT DISTINCT' => User::getTable() . '.id AS users_id',
-         'FIELDS'          => [User::getTable() . '.language AS language']
+         'FIELDS'          => [
+            User::getTable() . '.id AS users_id',
+            User::getTable() . '.language AS language'
+         ],
+         'DISTINCT'        => true,
       ];
    }