+++ inc/profile_user.class.php
@@ -450,13 +450,14 @@ class Profile_User extends CommonDBRelation {
       $putable = Profile_User::getTable();
       $etable = Entity::getTable();
       $iterator = $DB->request([
-         'SELECT DISTINCT' => "$utable.*",
-         'FIELDS'          => [
+         'SELECT'          => [
+            "$utable.*",
             "$putable.entities_id AS entity",
             "$putable.id AS linkID",
             "$putable.is_dynamic",
             "$putable.is_recursive"
          ],
+         'DISTINCT'        => true,
          'FROM'            => $putable,
          'LEFT JOIN'       => [
             $etable  => [
@@ -622,8 +623,11 @@ class Profile_User extends CommonDBRelation {
       global $DB;
 
       $iterator = $DB->request([
-         'SELECT DISTINCT' => 'entities_id',
-         'FIELDS'          => 'is_recursive',
+         'SELECT'          => [
+            'entities_id',
+            'is_recursive'
+         ],
+         'DISTINCT'        => true,
          'FROM'            => 'glpi_profiles_users',
          'WHERE'           => ['users_id' => $user_ID]
       ]);
@@ -674,8 +678,11 @@ class Profile_User extends CommonDBRelation {
       $ptable = Profile::getTable();
       $prtable = ProfileRight::getTable();
       $iterator = $DB->request([
-         'SELECT DISTINCT' => "$putable.entities_id",
-         'FIELDS'          => ["$putable.is_recursive"],
+         'SELECT'          => [
+            "$putable.entities_id",
+            "$putable.is_recursive"
+         ],
+         'DISTINCT'        => true,
          'FROM'            => $putable,
          'INNER JOIN'      => [
             $ptable  => [
@@ -738,7 +745,8 @@ class Profile_User extends CommonDBRelation {
       }
 
       $iterator = $DB->request([
-         'SELECT DISTINCT' => 'profiles_id',
+         'SELECT'          => 'profiles_id',
+         'DISTINCT'        => true,
          'FROM'            => 'glpi_profiles_users',
          'WHERE'           => $where
       ]);